Split the direction catalog by direction
catalog.ts mixed the shared interfaces, a 168-line array of 14 brand
directions, and the lookup. The data dominated the file and made it
impossible to format: expanding the compact objects takes it to 395
counted lines against a 300-line budget.

Moves the interfaces to brand-types.ts and each direction to its own file
under directions/, with an index that assembles DIRECTIONS. Consumers
import from the new locations rather than through a re-export, so there is
one path to each symbol.

No direction data changes. tsc is clean, all 29 tests pass, and every
direction still passes WCAG AA under cli check --all.
